Surface rendering: split SurfaceRenderer into orchestrator + helpers
SurfaceRenderer was a 1276-line god object handling shader init,
texture management, per-frame uniform writes, bind-group plumbing,
and the actual render-pass orchestration. Split into:

  SurfaceTextures (263 lines, new)
    Owns the six screen-sized textures + views; ensure/invalidate
    rebuild semantics tied to render-scale / water-quality changes.

  SurfaceShaders (552 lines, new)
    Owns the five shaders, samplers, uniform buffers, and bind groups.
    Hosts the per-frame update*Uniforms methods and the bind-group
    rebuild helpers.

  SurfaceRenderer (639 lines, was 1276)
    Now just the orchestrator: lifecycle, ensure-textures wrapper that
    coordinates with the wetness pipeline and bind-group invalidation,
    and onRender that reads as a clear sequence of passes.

Behavior is unchanged. A handful of smells noticed in the original
code (whenReady doesn't actually wait for shader compile errors;
hasTerrainData is dead-code-true; double getRenderer call) were left
in place — flagged for future work, not fixed in this refactor.
